Moved locales directory
