+++ internal/i18n/locales/fr-FR.yml
